refactor: move the bounds svg helpers into test fixtures
Per review -- the render/snapshot helpers live in tests/fixtures now
instead of inside the test file.
